requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-WTF==1.2.1
WTForms==3.1.2
email-validator==2.1.1
python-dotenv==1.0.1
PyMySQL==1.1.1
cryptography==42.0.5
Werkzeug==3.0.3
google-generativeai==0.8.3
Pillow==10.4.0
itsdangerous==2.2.0
